+++ spectify-frontend/src/app/(main)/following/more/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+import React from 'react'
+import { useState } from 'react';
+import FollowingComponent from "@/components/following/following";
+
+export default function MorePage() {
+    const [limit] = useState(Infinity);
+  return (
+    <>
+    <FollowingComponent limit={limit} />
+    </>
+  )
+}
